requirements.txt
python-binance
pandas
numpy
vectorbt
mplfinance
python-dotenv
requests
python-telegram-bot
binance-connector
binance
cryptography